refractor and changes
