--- src/main/java/net/rodofire/mushrooomsmod/feature/ModFeatures.java
@@ -1,7 +1,0 @@
-package net.rodofire.mushrooomsmod.feature;
-
-import net.minecraft.world.gen.feature.FeatureConfig;
-
-public abstract class ModFeatures <FC extends FeatureConfig>{
-
-}
