+++ src/main/java/net/rodofire/mushrooomsmod/world/tree/sapplinggenerator/BlueLuminescentSapplingGenerator.java
@@ -4,7 +4,7 @@ import net.minecraft.block.sapling.SaplingGenerator;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.gen.feature.ConfiguredFeature;
-import net.rodofire.mushrooomsmod.world.configuredfeatures.ModConfiguredFeatures;
+import net.rodofire.mushrooomsmod.world.features.configuredfeatures.ModConfiguredFeatures;
 import org.jetbrains.annotations.Nullable;
 
 public class BlueLuminescentSapplingGenerator extends SaplingGenerator {
